requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
python-dotenv>=1.0.0
redis>=4.0
requests>=2.28
